struct fuel now has a 'fuel' member of type double.

+++ xpilot/src/server/map.h
@@ -108,7 +108,7 @@
 
 struct fuel {
     clpos	pos;
-    long	fuel;
+    double	fuel;
     unsigned	conn_mask;
     long	last_change;
     int		team;
